feat(inventory): measure part stock in units, not just whole spares
Stock, reorder-at, restock quantity and the adjust_part_stock delta are all
decimal now, a part can say what unit it is counted in, and it can say how much
one completion draws down. Fabric softener topped up a third of a bottle at a
time stops reading as three bottles gone after three refills.

Quantities round to three places on every write and collapse back to an int when
whole, so 0.1 taken ten times lands on exactly zero and the ordinary
count-the-filters case still round-trips as 3 rather than 3.0. The two per-use
amounts have a read-side floor of 1, so parts written before the fields existed
keep consuming exactly one spare.

The unit follows the amount to every surface: the part chips, a linked task's
detail line, the stock number entity on the appliance device page, the low-stock
binary sensor attributes, and a new `unit` field on the three stock events. A
part that sets no unit and no per-use amount keeps stepping in whole spares, and
its stock box still refuses "2.5 filters".

(Fixes #220)

+++ custom_components/home_keeper/__init__.py
@@ -377,12 +377,14 @@ UPDATE_ASSET_DOCUMENT_SCHEMA = vol.Schema(
     }
 )
 
-# Adjust a part's on-hand spare count by a (signed) delta; clamped at zero.
+# Adjust a part's on-hand spare quantity by a (signed) delta; clamped at zero. The
+# delta is a float because stock is: a part measured in millilitres draws down by 250,
+# a bottle topped up a third at a time by 0.33.
 ADJUST_PART_STOCK_SCHEMA = vol.Schema(
     {
         vol.Required("asset_id"): cv.string,
         vol.Required("part_id"): cv.string,
-        vol.Required("delta"): vol.Coerce(int),
+        vol.Required("delta"): vol.Coerce(float),
     }
 )
 # Detach a part's attached file (upload is HTTP-only — see manuals.py — since a
